Db update

+++ models/EndUser.php
@@ -10,17 +10,23 @@
      * This is the model class for table "end_users".
      *
      * @property integer $id
+     * @property integer $distributor_id
      * @property integer $user_id
-     * @property string $title
-     * @property string $email
+     * @property string  $phone
+     * @property string  $contact_person
      * @property integer $country_id
-     *
-     * @property PurchaseOrder $purchaseOrder
-     * @property User $user
-     * @property Country $country
-     *
      * @property string $created_at
      * @property string $updated_at
+     *
+     * @property User $user
+     * @property string $title
+     * @property string $email
+     * @property PurchaseOrder[] $orders
+     * @property Country $country
+     * @property System[] $systems
+     * @property Distributor $distributor
+     *
+     *
      */
     class EndUser extends PpdBaseModel
     {
@@ -30,52 +36,6 @@
         public static function tableName()
         {
             return 'end_users';
-        }
-
-        /**
-         * @inheritdoc
-         */
-        public function rules()
-        {
-            return [
-                [['title', 'email', 'country_id'], 'required'],
-                [['title', 'email'], 'string', 'max' => 45],
-                [['user_id', 'country_id'], 'integer'],
-                [['email'], 'email'],
-                [['email'], 'unique'],
-            ];
-        }
-
-        public function fields()
-        {
-            return [
-                'id', 'title', 'email',
-                'country'    => function () {
-                        return (isset($this->country)) ? $this->country->name : null;
-                    },
-                'created_at' => function () {
-                        return date('M j Y h:i A', strtotime($this->created_at));
-                    },
-                'updated_at' => function () {
-                        return (!is_null($this->updated_at)) ? date('M j Y h:i A', strtotime($this->updated_at)) : null;
-                    }
-            ];
-        }
-
-        /**
-         * @inheritdoc
-         */
-        public function attributeLabels()
-        {
-            return [
-                'id'         => 'ID',
-                'title'      => Yii::t('app', 'Title'),
-                'email'      => Yii::t('app', 'Email'),
-                'systems'    => Yii::t('app', 'Systems'),
-                'country'    => Yii::t('app', 'Country'),
-                'user'       => Yii::t('app', 'User'),
-                'country_id' => Yii::t('app', 'Country'),
-            ];
         }
 
         /**
@@ -95,70 +55,103 @@
         }
 
         /**
-         * @return \yii\db\ActiveQuery
+         * @inheritdoc
          */
-        public function getSystems()
+        public function rules()
         {
-            return $this->hasMany(System::className(), ['end_user_id' => 'id']);
+            return [
+                [['user_id', 'country_id', 'distributor_id'], 'required'],
+                [['phone'], 'string', 'max' => 45],
+                [['contact_person'], 'string', 'max' => 128],
+                [['user_id', 'country_id', 'phone', 'created_at', 'updated_at', 'contact_person'], 'safe']
+            ];
         }
 
-        public function getPurchaseOrder()
+        public function fields()
         {
-            return $this->hasOne(PurchaseOrder::className(), ['end_user_id' => 'id']);
+            return [
+                'id', 'user_id', 'title', 'email', 'country_id', 'contact_person', 'country',
+                'created_at' => function () {
+                    return date('M j Y h:i A', strtotime($this->created_at));
+                },
+                'updated_at' => function () {
+                    return (!is_null($this->updated_at)) ? date('M j Y h:i A', strtotime($this->updated_at)) : null;
+                }
+            ];
+        }
+
+        public function extraFields()
+        {
+            return [
+                'user', 'orders', 'systems',
+            ];
         }
 
         /**
-         * Returns user account identity related to this end user record
-         *
-         * @return User|null
+         * @inheritdoc
          */
+        public function attributeLabels()
+        {
+            return [
+                'id'         => 'ID',
+                'title'      => Yii::t('app', 'Title'),
+                'email'      => Yii::t('app', 'Email'),
+                'systems'    => Yii::t('app', 'Systems'),
+                'orders'     => Yii::t('app', 'Purchase Orders'),
+                'country'    => Yii::t('app', 'Country'),
+                'user'       => Yii::t('app', 'User'),
+                'country_id' => Yii::t('app', 'Country'),
+                'distributor_id' => Yii::t('app', 'Distributor'),
+            ];
+        }
+
         public function getUser()
         {
             return $this->hasOne(User::className(), ['id' => 'user_id']);
         }
 
-        /**
-         * Returns country instances associated with this end user record
-         *
-         * @return Country
-         */
+        public function getTitle()
+        {
+            return $this->user->first_name . ' ' . $this->user->last_name;
+        }
+
+        public function getEmail()
+        {
+            return $this->user->email;
+        }
+
         public function getCountry()
         {
-            return $this->hasOne(Country::className(), ['id_countries' => 'country_id']);
+            return $this->hasOne(Country::className(), ['id' => 'country_id']);
+        }
+
+        public function getOrders()
+        {
+            return $this->hasMany(PurchaseOrder::className(), ['end_user_id' => 'id']);
+        }
+
+        public function getSystems()
+        {
+            $result = [];
+            foreach ($this->orders as $order) {
+                $result[] = $order->system;
+            }
+
+            return $result;
+        }
+
+        public function getDistributor()
+        {
+            return $this->hasOne(Distributor::className(), ['id' => 'distributor_id']);
         }
 
         /**
+         * @param bool $needValidate
          * @return boolean
          */
         public function createModel()
         {
-            if ($this->validate()) {
-                $user = new User();
-                $user->email = $this->email;
-                $user->first_name = $this->title;
-                $user->password = Yii::$app->security->generateRandomString(8);
-                $user->password_repeat = $user->password;
-                $user->roleField = User::ROLE_END_USER;
-
-                if ($user->createModel()) {
-                    $this->user_id = $user->id;
-                    if ($this->save()) {
-                        return true;
-                    } else {
-                        return false;
-                    }
-                } else {
-                    if (is_array($user->errors)) {
-                        foreach ($user->errors as $attribute => $error) {
-                            $this->addError($attribute, $error[0]);
-                        }
-                    }
-
-                    return false;
-                }
-            } else {
-                return false;
-            }
+            return $this->save();
         }
 
         /**
@@ -174,6 +167,10 @@
          */
         public function deleteModel()
         {
-            return $this->delete();
+            if ($this->user->delete()) {
+                return $this->delete();
+            } else {
+                return false;
+            }
         }
     }
